% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreditAnalysis.R
\docType{package}
\name{CreditAnalysis}
\alias{CreditAnalysis}
\alias{CreditAnalysis-package}
\title{Coursework for Audit Analytics II}
\description{
Documenting the coursework and final project for the Audit Analytics II
course at Rutgers University. This course was taken as part of the
Masters in Financial Accountancy program.
}
