% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditsales.R
\docType{data}
\name{creditsales}
\alias{creditsales}
\title{Credit Sales dataset}
\description{
Toy dataset used in weekly assigments for the Audit Analytics II course
at Rutgers University.
}
\keyword{data}
