% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorestimates.R
\docType{data}
\name{errorestimates}
\alias{errorestimates}
\title{Error estimates dataset}
\description{
TOy dataset used in weekly assigments for the Audit Analytics II course
at Rutgers University.
}
\keyword{data}
